move some items to config.ts
