+++ src/index.ts
@@ -1,4 +1,5 @@
 export { run, type Options } from './cli/run.js';
-export { Core, type Config, type ESLintOptions, configDefaults } from './core.js';
+export { Core } from './core.js';
+export { type Config, type ESLintOptions, configDefaults } from './config.js';
 export { takeRuleStatistics, type RuleStatistic } from './formatter/index.js';
 export { type FixableMaker, type SuggestionFilter, type FixContext } from './plugin/index.js';
